erika_cloud/pyproject.toml
[tool.poetry]
name = "erika-cloud"
version = "0.1.0"
description = ""
authors = ["Leonid Sokolov <leonid.sokolov@big-picture.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
bottle = "^0.12.25"
peewee = "^3.17.5"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
pytest = "^8.3.3"
pytest-django = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
